+++ ports/esp_idf/memfault/common/memfault_platform_demo_cli_cmds.c
@@ -66,8 +66,8 @@ static void IRAM_ATTR prv_timer_group0_isr(void *para) {
   // Always clear the interrupt:
   #if CONFIG_IDF_TARGET_ESP32
   TIMERG0.int_clr_timers.t0 = 1;
-  #elif CONFIG_IDF_TARGET_ESP32S2
-  TIMERG0.int_clr.t0 = 1;
+  #elif CONFIG_IDF_TARGET_ESP32S2 || CONFIG_IDF_TARGET_ESP32S3
+  TIMERG0.int_clr_timers.t0_int_clr = 1;
   #endif
 
   // Crash from ISR: