+++ tests/stubs/stub_assert.c
@@ -1,0 +1,10 @@
+//! @file
+//!
+//! Copyright (c) Memfault, Inc.
+//! See License.txt for details
+//!
+//! Stub function used to generate a non-empty archive to workaround the old
+//! circa-2005 'ar' utility installed by default on macs
+
+void stub_assert(void);
+void stub_assert(void) {}